# Avoid declare the exchange in every message sent