Avoid declare the exchange in every message sent also added benchmark test for EmmitAsync
result:
```
benchmark                old ns/op     new ns/op     delta
BenchmarkEmitAsync-4     326063        50410         -84.54%
```
resolve #3