@rafaeljesus 2017-08-01
@leandro-lugaresi Thank you mate, this request came from a couple of coworkers, this feature is pretty handy while creating services which only produces, but inevitably has the penalty you've just described